+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 - Generated typings for the `DeclarativeEventListeners` mixin are now available at `mixins/declarative-event-listeners.d.ts`.
 - [BREAKING] The `@observe` decorator now takes its dependencies as a rest parameter instead of an array (e.g. `@observer('foo', 'bar')` instead of `@observer(['foo', 'bar'])`.
+- [BREAKING] Decorators are now located at `polymer-decorators.js` instead of `global.js` (same for corresponding `.d.ts` file).
 
 ## [0.1.2] - 2018-01-01
 - Added `observer` and `computed` properties to the `@property` decorator options.